backend/requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.3
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary=1.44.2
colorama==0.4.6
Django==4.2.21
django-cors-headers==4.9.0
django-environ==0.13.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
gunicorn==25.3.0
kombu==5.6.2
mongoengine==0.29.3
packaging==26.1
pillow==12.2.0
prompt_toolkit==3.0.52
PyJWT==2.12.1
pymongo==4.16.0
python-dateutil==2.9.0.post0
redis==7.4.0
six==1.17.0
sqlparse==0.5.5
tzdata==2026.1
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
